backend/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# AI & Embeddings
openai==1.12.0
tiktoken==0.5.2
pinecone-client==3.0.0

# Document Processing
pypdf2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.3
httpx==0.26.0

# Connectors
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.116.0
slack-sdk==3.27.0
notion-client==2.2.1
msal==1.26.0

# Utilities
python-dotenv==1.0.1
pydantic==2.6.0
pydantic-settings==2.1.0
aiofiles==23.2.1
redis==5.0.1

# Development
pytest==8.0.0
pytest-asyncio==0.23.4
